refactor: remove core cursor, first pass
