+++ src/cmap/wire_protocol/query.ts
@@ -6,7 +6,7 @@ import { getReadPreference, isSharded, applyCommonQueryOptions } from './shared'
 import type { Document } from '../../bson';
 import type { Server } from '../../sdam/server';
 import type { ReadPreferenceLike } from '../../read_preference';
-import type { InternalCursorState } from '../../cursor/core_cursor';
+import type { InternalCursorState } from '../../cursor/cursor';
 
 /** @internal */
 export interface QueryOptions extends CommandOptions {
